+++ svws-core/src/main/java/de/svws_nrw/core/kursblockung/KursblockungMatrix.java
@@ -422,14 +422,16 @@ public class KursblockungMatrix {
 		return sb.toString();
 	}
 
-	/** Füllt die Matrix mit ganzzahligen zufälligen Zahlenwerten aus dem Intervall {@code [von;bis]}.
+	/**
+	 * Füllt die Matrix mit ganzzahligen zufälligen Zahlenwerten aus dem Intervall {@code [von;bis]}.
 	 *
 	 * @param von Der kleinstmögliche zufällige Wert (inklusive).
-	 * @param bis Der größtmögliche zufällige Wert (inklusive). */
+	 * @param bis Der größtmögliche zufällige Wert (inklusive).
+	 */
 	public void fuelleMitZufallszahlenVonBis(final int von, final int bis) {
 		for (int r = 0; r < rows; r++) {
 			for (int c = 0; c < cols; c++) {
-				matrix[r][c] = _random.nextInt(bis - von + 1) + von;
+				matrix[r][c] = _random.nextLong(bis - von + 1) + von;
 			}
 		}
 	}
